requirements.txt
dbt-bigquery==1.6.12
dbt-core==1.6.11
setuptools>=66.0.0,<67.0.0
google-cloud-storage==2.16.0
apache-airflow[celery]==2.9.0
apache-airflow-providers-google==10.1.0
